Refactor message rendering
