+++ apps/xmtp.chat/src/components/Messages/MessageContent.tsx
@@ -1,9 +1,5 @@
 import { Code } from "@mantine/core";
 import type { DecodedMessage } from "@xmtp/browser-sdk";
-import {
-  ContentTypeGroupUpdated,
-  type GroupUpdated,
-} from "@xmtp/content-type-group-updated";
 import { ContentTypeReply, type Reply } from "@xmtp/content-type-reply";
 import {
   ContentTypeTransactionReference,
@@ -14,11 +10,7 @@ import {
   type WalletSendCallsParams,
 } from "@xmtp/content-type-wallet-send-calls";
 import { FallbackContent } from "@/components/Messages/FallbackContent";
-import { GroupUpdatedContent } from "@/components/Messages/GroupUpdatedContent";
-import {
-  MessageContentWrapper,
-  type MessageContentAlign,
-} from "@/components/Messages/MessageContentWrapper";
+import { type MessageContentAlign } from "@/components/Messages/MessageContentWrapper";
 import { ReplyContent } from "@/components/Messages/ReplyContent";
 import { TextContent } from "@/components/Messages/TextContent";
 import { TransactionReferenceContent } from "@/components/Messages/TransactionReferenceContent";
@@ -26,101 +18,56 @@ import { WalletSendCallsContent } from "@/components/Messages/WalletSendCallsCon
 
 export type MessageContentProps = {
   align: MessageContentAlign;
-  senderInboxId: string;
-  message: DecodedMessage;
   scrollToMessage: (id: string) => void;
+  message: DecodedMessage;
 };
 
 export const MessageContent: React.FC<MessageContentProps> = ({
   message,
   align,
-  senderInboxId,
   scrollToMessage,
 }) => {
   if (message.contentType.sameAs(ContentTypeTransactionReference)) {
     return (
-      <MessageContentWrapper
-        align={align}
-        senderInboxId={senderInboxId}
-        sentAtNs={message.sentAtNs}>
-        <TransactionReferenceContent
-          content={message.content as TransactionReference}
-        />
-      </MessageContentWrapper>
+      <TransactionReferenceContent
+        content={message.content as TransactionReference}
+      />
     );
   }
 
   if (message.contentType.sameAs(ContentTypeWalletSendCalls)) {
     return (
-      <MessageContentWrapper
-        align={align}
-        senderInboxId={senderInboxId}
-        sentAtNs={message.sentAtNs}>
-        <WalletSendCallsContent
-          content={message.content as WalletSendCallsParams}
-          conversationId={message.conversationId}
-        />
-      </MessageContentWrapper>
-    );
-  }
-
-  if (message.contentType.sameAs(ContentTypeGroupUpdated)) {
-    return (
-      <GroupUpdatedContent
-        content={message.content as GroupUpdated}
-        sentAtNs={message.sentAtNs}
+      <WalletSendCallsContent
+        content={message.content as WalletSendCallsParams}
+        conversationId={message.conversationId}
       />
     );
   }
 
   if (message.contentType.sameAs(ContentTypeReply)) {
     return (
-      <MessageContentWrapper
+      <ReplyContent
         align={align}
-        senderInboxId={senderInboxId}
-        sentAtNs={message.sentAtNs}>
-        <ReplyContent
-          align={align}
-          message={message as DecodedMessage<Reply>}
-          scrollToMessage={scrollToMessage}
-        />
-      </MessageContentWrapper>
+        message={message as DecodedMessage<Reply>}
+        scrollToMessage={scrollToMessage}
+      />
     );
   }
 
   if (typeof message.content === "string") {
-    return (
-      <MessageContentWrapper
-        align={align}
-        senderInboxId={senderInboxId}
-        sentAtNs={message.sentAtNs}>
-        <TextContent text={message.content} />
-      </MessageContentWrapper>
-    );
+    return <TextContent text={message.content} />;
   }
 
   if (typeof message.fallback === "string") {
-    return (
-      <MessageContentWrapper
-        align={align}
-        senderInboxId={senderInboxId}
-        sentAtNs={message.sentAtNs}>
-        <FallbackContent text={message.fallback} />
-      </MessageContentWrapper>
-    );
+    return <FallbackContent text={message.fallback} />;
   }
 
   return (
-    <MessageContentWrapper
-      align={align}
-      senderInboxId={senderInboxId}
-      sentAtNs={message.sentAtNs}>
-      <Code
-        block
-        w="100%"
-        style={{ whiteSpace: "pre-wrap", wordBreak: "break-all" }}>
-        {JSON.stringify(message.content ?? message.fallback, null, 2)}
-      </Code>
-    </MessageContentWrapper>
+    <Code
+      block
+      w="100%"
+      style={{ whiteSpace: "pre-wrap", wordBreak: "break-all" }}>
+      {JSON.stringify(message.content ?? message.fallback, null, 2)}
+    </Code>
   );
 };
