+++ apps/xmtp.chat/src/components/Messages/ReplyContent.module.css
@@ -1,0 +1,4 @@
+.text {
+  text-decoration: underline;
+  cursor: pointer;
+}
